Add a sort function to the single year page
Added a custom select that sorts the list of groundhogs when it
changes. If there is No JS, the old link will still work.

I hadn't done up the boxes as list items, so changed the interface
on the homepage and on the years page to do that.

Also tweaked the input field for the groundhogs to make it a big
larger and hopefully easier to interact with.