Keep header on the index page and the logo on other pages
One of the bits of feedback that I got was that when you first load it on your
phone, you don't know what you're looking at.

Added back the full header on the index page and I'm now keeping the logo
around on subsequent pages.

Also, I added a tagline that hopefully explains what the site is about.