Add 2 new pages, return groundhogs by country of origin
Also update the API to accept query parameter for a "country" so that
we can return either Canadian or American groundhogs.

Also included some very brittle tests so that's cool.

Subpoint: remove the border around scrollable tables when you click them.
It was weird and doesn't need to be there.