Add a map svg to Map links
The map is kind of a major feature and it should be more prominent.
For now, this is okay, but it should maybe be even more called out.

Here's the link to the icon svg:
- https://thenounproject.com/icon/map-1024316/

Credit: Map by b farias from <a href="https://thenounproject.com/browse/icons/term/map/" target="_blank" title="Map Icons">Noun Project</a>