Add retirement and successor data to groundhog profile page
Instead of saying "didn't make a prediction", say that the groundhog
retired and link to the new groundhog.

There's definitely some work to be done here to generalize the
message (right now it is mostly hardcoded), but the basics are in place.

Next year, we can fix it (maybe), depending on what happens.